Add per compiler supported wmma combinations
