+++ crates/cubecl-cpp/src/cuda/arch.rs
@@ -1,0 +1,3 @@
+pub struct CudaArchitecture {
+    pub version: u32,
+}
