+++ crates/cubecl-cpp/src/cuda/wmma/mod.rs
@@ -1,0 +1,2 @@
+pub mod cuda_compiler;
+pub use cuda_compiler::*;
