+++ crates/cubecl-cpp/src/hip/wmma/mod.rs
@@ -1,0 +1,4 @@
+pub mod rocwmma_compiler;
+pub use rocwmma_compiler::*;
+pub mod intrinsic_compiler;
+pub use intrinsic_compiler::*;
